# Conflicts:
#	.idea/vcs.xml
#	.idea/workspace.xml
#	eresViewer/.idea/vcs.xml

#	eresViewer/app/src/main/java/com/example/maciek/eresviewer/MainActivity.java
#	eresViewer/app/src/main/res/layout/activity_main.xml
#	eresViewer/app/src/main/res/layout/activity_subject.xml

+++ eresViewer/app/src/main/java/com/example/maciek/eresviewer/Mark.java
@@ -28,7 +28,7 @@ public class Mark {
     /**
      * Liczba ocen
      */
-    private float amountOfMarks;
+    private int amountOfMarks;
 
    //gettery
 
@@ -36,10 +36,10 @@ public class Mark {
     public float getMyMark() { return myMark; }
     public float getAverageMark() {return averageMark; }
     public float getHigherMark() { return higherMark; }
-    public float getAmountOfMarks() { return amountOfMarks; }
+    public int getAmountOfMarks() { return amountOfMarks; }
     public float getLowerMark() { return lowerMark; }
 
-    Mark(String markTitle, float myMark, float lowerMark,float averageMark,float higherMark, float amountOfMarks){
+    Mark(String markTitle, float myMark, float lowerMark,float averageMark,float higherMark, int amountOfMarks){
         this.markTitle=markTitle;
         this.myMark=myMark;
         this.lowerMark=lowerMark;
